+++ odoo/local-src/specific_crm/models/__init__.py
@@ -1,3 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from . import crm_lead
+from . import crm_industry
+from . import crm_lead_quicktitle
+from . import crm_lead2opportunity
+from . import res_partner