+++ odoo/local-src/specific_crm/models/crm_lead_quicktitle.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+from odoo import models, fields
+
+
+class CrmLeadQuicktitle(models.Model):
+    _name = "crm.lead.quicktitle"
+
+    name = fields.Char(
+        string="Name",
+        required=True,
+    )